back_end/login/requirements.txt
Django==4.0.3
djangorestframework==3.14.0
mysql-connector-python==8.0.26
pika==1.3.2
mysqlclient==2.2.0
django-mysql==4.10.0
django-cors-headers==4.1.0
